+++ tests/Feature/Framework/UsingConfigWithBootstrapSetup/FeatureTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\UsingConfigWithBootstrapSetup;
 
+use Gacela\Framework\Bootstrap\SetupGacela;
 use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 use Gacela\Framework\Gacela;
-use Gacela\Framework\Setup\SetupGacela;
 use PHPUnit\Framework\TestCase;
 
 final class FeatureTest extends TestCase
@@ -14,7 +14,7 @@ final class FeatureTest extends TestCase
     public function setUp(): void
     {
         $setup = (new SetupGacela())
-            ->setConfig(static function (ConfigBuilder $configBuilder): void {
+            ->setConfigFn(static function (ConfigBuilder $configBuilder): void {
                 $configBuilder->add('custom-config.php', 'custom-config_local.php');
             });