+++ tests/Feature/Framework/UsingCustomSuffixTypes/FeatureTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\UsingCustomSuffixTypes;
 
+use Gacela\Framework\Bootstrap\SetupGacela;
 use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 use Gacela\Framework\Gacela;
-use Gacela\Framework\Setup\SetupGacela;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -17,7 +17,7 @@ final class FeatureTest extends TestCase
     protected function setUp(): void
     {
         $setup = (new SetupGacela())
-            ->setConfig(static function (ConfigBuilder $configBuilder): void {
+            ->setConfigFn(static function (ConfigBuilder $configBuilder): void {
                 $configBuilder->add('config/*.php', 'config/local.php');
             });