+++ tests/Feature/Framework/UsingCustomSuffixTypes/gacela.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\UsingCustomSuffixTypes;
 
+use Gacela\Framework\Bootstrap\SetupGacela;
 use Gacela\Framework\Config\GacelaConfigBuilder\SuffixTypesBuilder;
-use Gacela\Framework\Setup\SetupGacela;
 
-return static fn () => (new SetupGacela())
-    ->setSuffixTypes(
+return (new SetupGacela())
+    ->setSuffixTypesFn(
         static function (SuffixTypesBuilder $suffixTypesBuilder): void {
             $suffixTypesBuilder
                 ->addFacade('FacaModuleA')